requirements.txt
Adafruit-Blinka==8.30.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-framebuf==1.6.5
adafruit-circuitpython-requests==2.0.4
adafruit-circuitpython-ssd1306==2.12.16
adafruit-circuitpython-typing==1.9.6
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.58.0
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==5.0.1
django-redis==5.4.0
docopt==0.6.2
h5py==3.10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
numpy==1.26.3
pandas==2.1.4
pillow==10.2.0
pipreqs==0.4.13
public-ip==0.12
pvlib==0.10.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyftdi==0.55.0
pynmeagps==1.0.33
pyOpenSSL==23.3.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
pyudev==0.24.1
pyusb==1.2.1
pyzmq==25.1.2
redis==5.0.1
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
RPIO==0.10.0
scipy==1.11.4
service-identity==24.1.0
six==1.16.0
spidev==3.6
sqlparse==0.4.4
sysv-ipc==1.1.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
yarg==0.1.9
zmq==0.0.0
zope.interface==6.1
django-bootstrap-v5